Added Product Class
