+++ GrabHub/AddProduct.cs
@@ -10,20 +10,7 @@ using System.Windows.Forms;
 
 namespace GrabHub
 {
-    public class Product
-    {
-        public string Name { get; set; }
-        public double Price { get; set; }
-        public int Quantity { get; set; }
-
-        public Product(string name, double price, int quantity)
-        {
-            Name = name;
-            Price = price;
-            Quantity = quantity;
-        }
-    }
-
+    
     public partial class AddProduct : Form
     {
         public static List<Product> orderedProducts = new List<Product>();
